backend/requirements.txt
fastapi==0.95.0
psycopg2-binary==2.9.10
uvicorn==0.22.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv
sqlalchemy
pydantic[email]
python-multipart
httpx
authlib
pinecone 
langchain
langchain-community
langchain-pinecone 
langchain-openai 
langchain-text-splitters 
